Refactor omnisci_xxx to heavyai/xxx
