+++ rbc/heavyai/__init__.py
@@ -1,11 +1,11 @@
-from .omnisci_array import *  # noqa: F401, F403
-from .omnisci_column import *  # noqa: F401, F403
-from .omnisci_bytes import *  # noqa: F401, F403
-from .omnisci_metatype import *  # noqa: F401, F403
-from .omnisci_text_encoding import *  # noqa: F401, F403
-from .omnisci_pipeline import *  # noqa: F401, F403
-from .omnisci_column_list import *  # noqa: F401, F403
-from .omnisci_table_function_manager import *  # noqa: F401, F403
+from .array import *  # noqa: F401, F403
+from .column import *  # noqa: F401, F403
+from .bytes import *  # noqa: F401, F403
+from .metatype import *  # noqa: F401, F403
+from .text_encoding import *  # noqa: F401, F403
+from .pipeline import *  # noqa: F401, F403
+from .column_list import *  # noqa: F401, F403
+from .table_function_manager import *  # noqa: F401, F403
 
 import rbc.heavyai.mathimpl as math  # noqa: F401
 import rbc.heavyai.npyimpl as np  # noqa: F401
