+++ rbc/heavyai/column.py
@@ -9,8 +9,8 @@ __all__ = ['OutputColumn', 'Column', 'OmnisciOutputColumnType', 'OmnisciColumnTy
 
 from llvmlite import ir
 from rbc import typesystem, irutils
-from .omnisci_buffer import Buffer, OmnisciBufferType, BufferType
-from .omnisci_column_list import OmnisciColumnListType
+from .buffer import Buffer, OmnisciBufferType, BufferType
+from .column_list import OmnisciColumnListType
 from rbc.targetinfo import TargetInfo
 from numba.core import extending, types
