+++ rbc/heavyai/pipeline.py
@@ -1,7 +1,7 @@
 import operator
 
 from rbc.errors import NumbaTypeError
-from .omnisci_buffer import BufferMeta, free_all_other_buffers
+from .buffer import BufferMeta, free_all_other_buffers
 from numba.core import ir, types
 from numba.core.compiler import CompilerBase, DefaultPassBuilder
 from numba.core.compiler_machinery import FunctionPass, register_pass
