+++ rbc/heavyai/table_function_manager.py
