+++ rbc/tests/heavyai/test_device_selection.py
